move check_empty decorator to helpers module
